Support wakeup device via send power key
It's hard to find out the power key define in the device
and the right input device for it, but it worth to do.

Monitor device screen off status, and click on screen
to wake it up. so Cool! also do not capture screen when
it's off.

Fixed an ugly call setZvalue for grayMask cause prompt
text grayed out.