Update screen ASAP when try to wakeup device
So that we can see th lock screen as soon as possible,
and user will not try to click on the old screen which
is already passed.

Don't hide gray mask until new frame available

So that do not give user an illusion of device is clickable
now, and do not show the passed screen content to user.